+++ angular/lib/src/compiler/html_ast.dart
@@ -22,10 +22,14 @@ class HtmlTextAst implements HtmlAst {
 class HtmlAttrAst implements HtmlAst {
   final String name;
   final String value;
+
   @override
   final SourceSpan sourceSpan;
 
-  HtmlAttrAst(this.name, this.value, this.sourceSpan);
+  /// True if this attribute has an explicit value.
+  final bool hasValue;
+
+  HtmlAttrAst(this.name, this.value, this.sourceSpan, this.hasValue);
 
   @override
   visit(HtmlAstVisitor visitor, dynamic context) {
@@ -88,7 +92,7 @@ List htmlVisitAll(
   context,
 ]) {
   var result = [];
-  asts.forEach((ast) {
+  for (var ast in asts) {
     bool handled = visitor.visit(ast, context);
     if (!handled) {
       var astResult = ast.visit(visitor, context);
@@ -96,6 +100,6 @@ List htmlVisitAll(
         result.add(astResult);
       }
     }
-  });
+  }
   return result;
 }