+++ angular/lib/src/source_gen/common/reflection_info_model.dart
@@ -60,13 +60,8 @@ class ReflectionInfoModel {
       _allAnnotations,
       _parameterList
     ];
-
     if (!isFunction) {
       reflectionInfoParams.add(_factoryClosure);
-
-      if (_interfaces.isNotEmpty) {
-        reflectionInfoParams.add(_interfaceList);
-      }
     }
     return reflectionInfoParams;
   }
@@ -81,14 +76,11 @@ class ReflectionInfoModel {
       list(_parameters.map((ParameterModel model) => model.asList),
           asConst: true);
 
-  ExpressionBuilder get _interfaceList =>
-      list(_interfaces, type: lib$core.$dynamic, asConst: true);
-
   ExpressionBuilder get _factoryClosure {
     var closure = new MethodBuilder.closure(returns: _constructorExpression);
-    _parameters.forEach((param) {
+    for (var param in _parameters) {
       closure.addPositional(param.asBuilder);
-    });
+    }
     return closure;
   }