+++ angular/lib/src/transform/transformer.dart
@@ -1,20 +1,16 @@
 import 'package:angular_compiler/angular_compiler.dart';
 import 'package:barback/barback.dart';
 import 'package:build_barback/build_barback.dart';
-import 'package:dart_style/dart_style.dart';
 
 import '../../source_gen.dart';
 import 'common/eager_transformer_wrapper.dart';
-import 'common/formatter.dart' as formatter;
 import 'deferred_rewriter/transformer.dart';
 import 'reflection_remover/transformer.dart';
 import 'stylesheet_compiler/transformer.dart';
 
 /// Replaces Angular 2 mirror use with generated code.
 class AngularTransformerGroup extends TransformerGroup {
-  AngularTransformerGroup._(Iterable<Iterable> phases) : super(phases) {
-    formatter.init(new DartFormatter());
-  }
+  AngularTransformerGroup._(Iterable<Iterable> phases) : super(phases);
 
   factory AngularTransformerGroup(CompilerFlags flags) {
     Iterable<Iterable> phases = [